update the firestore-client with datastore from client
